+++ src/GafferScene/InteractiveRender.cpp
@@ -148,18 +148,13 @@ size_t InteractiveRender::g_firstPlugIndex = 0;
 GAFFER_NODE_DEFINE_TYPE( InteractiveRender );
 
 InteractiveRender::InteractiveRender( const std::string &name )
-	:	InteractiveRender( /* rendererType = */ InternedString(), name )
-{
-}
-
-InteractiveRender::InteractiveRender( const IECore::InternedString &rendererType, const std::string &name )
 	:	ComputeNode( name ),
 		m_state( Stopped ),
 		m_messageHandler( new RenderMessageHandler() )
 {
 	storeIndexOfNextChild( g_firstPlugIndex );
 	addChild( new ScenePlug( "in" ) );
-	addChild( new StringPlug( rendererType.string().empty() ? "renderer" : "__renderer", Plug::In, rendererType.string() ) );
+	addChild( new StringPlug( "renderer" ) );
 	addChild( new IntPlug( "state", Plug::In, Stopped, Stopped, Paused, Plug::Default & ~Plug::Serialisable ) );
 	addChild( new ScenePlug( "out", Plug::Out, Plug::Default & ~Plug::Serialisable ) );
 	addChild( new StringPlug( "resolvedRenderer", Plug::Out, "", Plug::Default & ~Plug::Serialisable ) );