+++ Changes.md
@@ -44,6 +44,7 @@ Breaking Changes
 - InteractiveRenderTest : Removed `interactiveRenderNodeClass` member and `useNodeClass` argument to `_createInteractiveRender`. All testing is now performed with the InteractiveRender node itself.
 - GafferCycles : The `devices`, `nodes`, `shaders`, `lights`, and `passes` Python attributes now contain IECore.CompoundData instead of Python dictionaries.
 - InteractiveArnoldRender, InteractiveCyclesRender, InteractiveDelightRender : Removed. Use the generic InteractiveRender node instead.
+- InteractiveRender : Removed protected constructor for creating renderer-specific derived classes.
 
 1.5.0.0a3 (relative to 1.5.0.0a2)
 =========